Enhance ACPI support to bring up all X2APIC cpus and memory plus some small fixes
- Add detection for X2APIC CPUs in ACPI startup
  (makes visible CPUs with ids>=256)
- Add detection for memory affinity for X2APIC CPUs in ACPI
  startup (binds CPUs with ids>=256 to memory regions)
- Minor tweaks to reduce non-debugging output on
  startup on machines with many cpus

+++ include/nautilus/smp.h
@@ -74,7 +74,7 @@ struct cpu {
     struct nk_thread * cur_thread; /* KCH: this must be first! */
 
     cpu_id_t id;
-    uint8_t lapic_id;
+    uint32_t lapic_id;   
     uint8_t enabled;
     uint8_t is_bsp;
     uint32_t cpu_sig;
